+++ test/fixtures/send-data-process/return-data.js
@@ -1,0 +1,11 @@
+
+process.on('message', function(packet) {
+  if (packet.type == 'process:msg') {
+    process.send({
+      type : 'process:msg',
+      data : {
+        success : true
+      }
+    });
+  }
+});